remove unneccessary fields on MetaKGDoc; move some methods which only used by SmartAPIEntity to SmartAPIEntity
