+++ src/controller/base.py
@@ -8,8 +8,8 @@ import jsonschema
 from configparser import ConfigParser
 from elasticsearch.exceptions import NotFoundError as ESNotFoundError
 
-from utils import decoder, monitor
-from utils.downloader import download, Downloader, File
+from utils import decoder
+from utils.downloader import Downloader, File
 from .exceptions import ControllerError, NotFoundError
 
 # NOTE
@@ -121,11 +121,6 @@ class AbstractWebEntity(ABC):
         assert urlparse(str(url)).scheme in ("http", "https")
 
         self._url = url
-
-        self.uptime = APIMonitorStatus(self)
-        self.webdoc = APIRefreshStatus(self)
-
-        self._raw = None
         self._data = {}
 
     @staticmethod
@@ -201,28 +196,7 @@ class AbstractWebEntity(ABC):
             raise NotFoundError from err
         obj = cls(doc.get_url())
         obj._doc = doc
-        obj.raw = decoder.decompress(doc._raw)
-
-        obj.uptime = APIMonitorStatus(
-            obj,
-            (
-                doc._status.uptime_status,
-                doc._status.uptime_msg,
-            ),
-            doc._status.uptime_ts,
-        )
-
-        obj.webdoc = APIRefreshStatus(
-            obj,
-            doc._status.refresh_status,
-            doc._status.refresh_ts,
-        )
-
         return obj
-
-    @classmethod
-    def get_doc_url(cls, doc):
-        raise NotImplementedError()
 
     @property
     def _id(self):
@@ -242,31 +216,6 @@ class AbstractWebEntity(ABC):
         # it's not trivial to handle url change in DB
         return self._url
 
-    @property
-    def raw(self):
-        """
-        Bytes that correspond to the URL.
-        This object is a view of this field.
-        """
-        return self._raw
-
-    @raw.setter
-    def raw(self, value):
-
-        if not value:
-            raise ControllerError("Empty value.")
-        try:
-            self._data = decoder.to_dict(value)
-        except (ValueError, TypeError) as err:
-            raise ControllerError(str(err)) from err
-        else:  # dict conversion success
-            self._raw = value
-
-        # update the timestamps
-        self.last_updated = datetime.now(timezone.utc)
-        if hasattr(self, 'date_created') and not self.date_created:
-            self.date_created = self.last_updated
-
     def _validate_dispatch(self):
         raise NotImplementedError()
 
@@ -285,17 +234,6 @@ class AbstractWebEntity(ABC):
 
         return doc
 
-    def check(self):
-
-        doc = dict(self)
-        doc["_id"] = self._id
-
-        api = monitor.API(doc)
-        api.check_api_status()  # blocking network operation
-        status = api.get_api_status()
-        self.uptime.update(status)
-        return status
-
     def delete(self):
         try:
             self.MODEL_CLASS.get(self._id).delete()
@@ -304,48 +242,7 @@ class AbstractWebEntity(ABC):
 
         return self._id
 
-    def refresh(self, file=None):
-
-        if file is None:  # blocking network operation
-            file = download(self.url, raise_error=False)
-
-        self.webdoc.update(file)
-        return self.webdoc.status
-
-    def save(self, force_save=True):
-        # TODO DOCSTRING
-
-        if not self.raw:
-            raise ControllerError("No content.")
-
-        if self.url is self.VALIDATION_ONLY:
-            raise ControllerError("In validation-only mode.")
-
-        # NOTE
-        # why not enforce validation here?
-        # we add additional constraints to the application from time to time
-        # it's actually hard to retrospectively make sure all previously
-        # submitted API document always meet our latest requirements
-
-        _doc = self._validate_dispatch()
-        _doc.transform()
-
-        doc = self.MODEL_CLASS(**_doc)
-
-        if self.uptime.status:
-            doc._status.uptime_status = self.uptime.status[0]
-            doc._status.uptime_msg = self.uptime.status[1]
-        doc._status.uptime_ts = self.uptime.timestamp
-
-        doc._status.refresh_status = self.webdoc.status
-        doc._status.refresh_ts = self.webdoc.timestamp
-
-        doc._raw = decoder.compress(self.raw)
-
-        if force_save:
-            doc.save(skip_empty=False)
-
-        self._doc = doc
+    def save(self):
         return self._id
 
     # READ-ONLY DICT-LIKE ACCESS
