+++ src/controller/metakg.py
@@ -1,40 +1,4 @@
-"""
-    MetaKG CRUD Validation and Refresh Operations
-
-    Validation only:
-
-        meta_kg = MetaKG(MetaKG.VALIDATION_ONLY)
-        meta_kg.raw = rawbytes
-        meta_kg.validate()
-
-    Add a document:
-
-        meta_kg = MetaKG(url)
-        meta_kg.raw = rawbytes
-
-        meta_kg.username = username
-
-        meta_kg.validate() # should be called before saving
-        meta_kg.save()
-
-        meta_kg.check() # populate uptime status
-        meta_kg.refresh() # refresh and populate refresh status, auto validate
-        meta_kg.save()
-
-    Modify a document metadata:
-
-        meta_kg = MetaKG.get(_id)
-        meta_kg.slug = newslug
-        meta_kg.save()
-
-    Delete a document:
-
-        meta_kg = MetaKG.get(_id)
-        meta_kg.delete()
-
-"""
 import logging
-import json
 from collections.abc import Mapping
 
 from model import MetaKGDoc
@@ -67,5 +31,4 @@ class MetaKGEntity(AbstractWebEntity, Mapping):
     @classmethod
     def create_doc(cls, metadata):
         doc = MetaKGDoc(**metadata)
-        doc._raw = decoder.compress(json.dumps(metadata).encode())
         doc.save()
